Move reminder check to reminders cog
